% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ex_metagenome_input}
\alias{ex_metagenome_input}
\title{An example dataset to be used with digestR}
\format{A wide-format data frame with some columns and some rows}
\usage{
data(ex_metagenome_input)
}
\description{
This data frame is some stuff
}
\keyword{datasets}

