% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_signif}
\alias{find_signif}
\title{Finds significant trends}
\usage{
find_signif(d, alpha)
}
\arguments{
\item{d}{data frame for which slopes have been calculated}

\item{alpha}{alpha value; defaults to 0.05}

\item{indep.var}{independent variable, as a character}

\item{dep.var}{}
}
\description{
This is a bit of a placeholder function - later I will come up with a more statistically valid way of finding true trends
}

