% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_trend}
\alias{find_trend}
\title{Quantifies linear trends in pfam or fxn with independant variable}
\usage{
find_trend(d, indep.var = "depth", param = "frac.reads", disc = "pfam")
}
\description{
Quantifies linear trends in pfam or fxn with independant variable
}

