\name{fo}
\alias{fo}
\title{Description of the fo function}
\usage{
fo(your_arguments_here)
}
\description{
Provide a brief description of what the function 'fo' does.
}
\arguments{
  \item{your_argument_name}{Description of the argument.}
  ...
}
\details{
More detailed explanation of the function, if necessary.
}
\examples{
fo(your_example_arguments)
}
\value{
Description of the value returned by the function.
}
\seealso{
Any related functions or references.
}
