% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_signif}
\alias{plot_signif}
\title{Plots significant reads}
\usage{
plot_signif(d, indep.var = "depth", dep.var = "frac.reads",
  variable = "pfam", alpha = 0.05)
}
\arguments{
\item{d}{a data frame created by read_digestR}

\item{indep.var}{character}

\item{dep.var}{character}

\item{variable}{variable by which to to facet the data}

\item{alpha}{significance cutoff}
}
\description{
Plots significant reads
}

