% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_digestR}
\alias{read_digestR}
\title{Reads the 3 necessary files for digestR analysis}
\usage{
read_digestR(data.fn = NULL, key.fn = NULL, fxn.map.fn = NULL,
  normalizer = "sum.reads", elim.zeroes = TRUE)
}
\arguments{
\item{data.fn}{}

\item{key.fn}{}

\item{fxn.map.fn}{}
}
\description{
digestR requires a dataset, a key, and a map. This function reads all of these either from user-supplied files or default examples.
}

